--- src/constants/interface/comment.interface.ts
@@ -1,14 +1,0 @@
-import firebase from 'firebase/app'
-
-export interface Comment {
-  commentId: string
-  lectureId: string
-  userId: string
-  message: string
-  createAt: firebase.firestore.Timestamp
-  updateAt: firebase.firestore.Timestamp
-
-  //-----------------------
-  displayName: string
-  imageURL: string
-}
