+++ src/constants/interface/lecture.interface.ts
@@ -54,3 +54,14 @@ export interface QAndA {
   createAt?: firebase.firestore.Timestamp
   updateAt?: firebase.firestore.Timestamp
 }
+
+export interface subject {
+  subjectId: string
+  subjectCode: string
+  subjectName: string
+  subjectYear: string
+  subjectGroup: string
+  subjectGroup2?: string
+  subjectFaculty: string
+  subjectMajor: string
+}
