--- src/constants/interface/subject.interface.ts
@@ -1,10 +1,0 @@
-export interface subject {
-  subjectId: string
-  subjectCode: string
-  subjectName: string
-  subjectYear: string
-  subjectGroup: string
-  subjectGroup2?: string
-  subjectFaculty: string
-  subjectMajor: string
-}
